# kanban-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kanban Board API",
        "description": "API para gestionar Tableros, Tarjetas y Tareas de un sistema Kanban, siguiendo la especificación OpenAPI 3.",
        "contact": {
            "name": "Tu Nombre",
            "email": "tu.email@ejemplo.com"
        },
        "license": {
            "name": "Licencia Estándar de API",
            "url": "http://localhost/licencia"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtiene todas las tareas",
                "description": "Recupera una lista de todas las tareas existentes.",
                "operationId": "4293c00a4cffdb881d9f6c2099875f96",
                "responses": {
                    "200": {
                        "description": "Lista de tareas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Task"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Crea una nueva tarea",
                "operationId": "379ae513b47f83882f91f2b9143fc425",
                "requestBody": {
                    "description": "Datos de la tarea a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "taskName",
                                    "description",
                                    "startDate",
                                    "endDate",
                                    "allocator",
                                    "priority",
                                    "progress",
                                    "card_id"
                                ],
                                "properties": {
                                    "taskName": {
                                        "type": "string",
                                        "example": "Implementar OAuth"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Configurar la autenticación con Google y Facebook."
                                    },
                                    "startDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-01"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-15"
                                    },
                                    "allocator": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "employee": {
                                        "type": "string",
                                        "example": "John Smith",
                                        "nullable": true
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "Alta",
                                            "Media",
                                            "Baja"
                                        ],
                                        "example": "Alta"
                                    },
                                    "progress": {
                                        "type": "string",
                                        "example": "Pendiente"
                                    },
                                    "card_id": {
                                        "type": "integer",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/v1/tasks/{id}": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtiene una tarea específica por ID",
                "operationId": "5c89555764433d66ee2d4a8aefc4b0c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a buscar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Actualiza los detalles de una tarea",
                "description": "Permite actualizar el nombre, descripción, prioridad y/o fecha de finalización de una tarea.",
                "operationId": "2d67628d964268e2b57c176e100ccf23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Campos de la tarea a actualizar",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "taskName": {
                                        "type": "string",
                                        "example": "Implementar Autenticación"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada."
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "Alta",
                                            "Media",
                                            "Baja"
                                        ],
                                        "example": "Media"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-20",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Elimina una tarea por su ID",
                "operationId": "4b08d06f1cae255ad9f54935cc9d8fed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarea eliminada exitosamente (No Content)"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            }
        },
        "/api/v1/tasks/{id}/move": {
            "patch": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Mueve una tarea a una nueva tarjeta (columna)",
                "description": "Método utilizado para cambiar la tarjeta (columna) a la que pertenece la tarea (ej. de 'Por Hacer' a 'En Progreso').",
                "operationId": "313340822475231c04ec3f0dbc73c376",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea a mover",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Nuevo ID de la tarjeta de destino",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_id"
                                ],
                                "properties": {
                                    "card_id": {
                                        "description": "El ID de la tarjeta a la que se moverá la tarea.",
                                        "type": "integer",
                                        "example": "8"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ID de tarjeta actualizado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            }
        },
        "/api/v1/tasks/{id}/days-left": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Calcula los días restantes hasta la fecha de vencimiento",
                "description": "Devuelve el número de días que faltan para el endDate de la tarea.",
                "operationId": "6f3acff214a7ee240f963bae81033332",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Días restantes calculados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "taskName": {
                                            "type": "string",
                                            "example": "Implementar OAuth"
                                        },
                                        "endDate": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-10-15"
                                        },
                                        "remaining_days": {
                                            "description": "Días restantes. Negativo si la fecha ya pasó.",
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "detail": {
                                            "type": "string",
                                            "example": "Queda tiempo o La tarea ya termino"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Task": {
                "title": "Task",
                "description": "Modelo de datos de una Tarea, el elemento individual dentro de una Tarjeta (Card).",
                "properties": {
                    "id": {
                        "description": "ID único de la tarea",
                        "type": "integer"
                    },
                    "card_id": {
                        "description": "ID de la tarjeta (columna) a la que pertenece la tarea",
                        "type": "integer"
                    },
                    "taskName": {
                        "description": "Nombre o título de la tarea",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descripción detallada de la tarea",
                        "type": "string"
                    },
                    "startDate": {
                        "description": "Fecha de inicio de la tarea",
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "description": "Fecha de vencimiento o fin de la tarea",
                        "type": "string",
                        "format": "date"
                    },
                    "allocator": {
                        "description": "Usuario que asignó la tarea",
                        "type": "string"
                    },
                    "employee": {
                        "description": "Usuario al que está asignada la tarea",
                        "type": "string",
                        "nullable": true
                    },
                    "priority": {
                        "description": "Prioridad de la tarea",
                        "type": "string",
                        "enum": [
                            "Alta",
                            "Media",
                            "Baja"
                        ]
                    },
                    "progress": {
                        "description": "Estado del progreso (ej. Pendiente, En curso, Bloqueado)",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}